Extract locales list and validation from resolver
